Start work on a new command system.
The current command system is all API implemented, and uses final classes. It is very difficult to change without breaking, so instead, this is an entirely new system built to be Common implemented.

Currently, most of the work has focused on parameters. Instead of wrapping parameters/arguments, we now modularise the code and use a builder to put it all together, a bit like CommandSpec.

A specification will go up on the PR once I get around to putting one up.